Fixed some bugs in Descriptors based classes, added UBOs and a camera

+++ Src/Vk/Model/Buffers/Buffer.cpp
@@ -2,17 +2,70 @@
 #include "../../Vendor/vma/vk_mem_alloc.h"
 #include "Buffer.h"
 
+#include "vulkan/vulkan_core.h"
+#include "vulkan/vulkan_enums.hpp"
 #include "vulkan/vulkan_structs.hpp"
 #include "../Services/ServiceLocator.h"
+#include "../../../Log/Log.h"
 #include <alloca.h>
 #include <cstdint>
+#include <stdexcept>
 
 namespace VkCore
 {
 
+    Buffer& Buffer::operator=(Buffer&& other)
+    {
+        if (this != &other)
+        {
+            m_Size = other.m_Size;
+            other.m_Size = 0;
+
+            m_UsageFlags = other.m_UsageFlags;
+            m_IsHostVisible = other.m_IsHostVisible;
+            m_IsMapped = other.m_IsMapped;
+
+            m_Buffer = other.m_Buffer;
+            other.m_Buffer = VK_NULL_HANDLE;
+
+            m_Allocation = other.m_Allocation;
+            other.m_Allocation = VK_NULL_HANDLE;
+
+            m_AllocationInfo = other.m_AllocationInfo;
+            other.m_AllocationInfo = {};
+        }
+
+        return *this;
+    }
+
+    Buffer::Buffer(Buffer&& other)
+    {
+        if (this != &other)
+        {
+            m_Size = other.m_Size;
+            other.m_Size = 0;
+
+            m_UsageFlags = other.m_UsageFlags;
+            m_IsHostVisible = other.m_IsHostVisible;
+            m_IsMapped = other.m_IsMapped;
+
+            m_Buffer = other.m_Buffer;
+            other.m_Buffer = VK_NULL_HANDLE;
+
+            m_Allocation = other.m_Allocation;
+            other.m_Allocation = VK_NULL_HANDLE;
+
+            m_AllocationInfo = other.m_AllocationInfo;
+            other.m_AllocationInfo = {};
+        }
+    }
+
     Buffer::~Buffer()
     {
-        ServiceLocator::GetAllocatorService().DestroyBuffer(*this);
+        if (m_Buffer != VK_NULL_HANDLE)
+        {
+            ServiceLocator::GetAllocatorService().DestroyBuffer(*this);
+        }
     }
 
     void Buffer::InitializeOnGpu(const void* data, const size_t size)
@@ -24,9 +77,17 @@ namespace VkCore
 
         m_Buffer =
             ServiceLocator::GetAllocatorService().CreateBufferOnGpu(data, bufferInfo, m_Allocation, &m_AllocationInfo);
+
+        m_Size = size;
     }
 
-    void Buffer::InitializeOnCpu(const void* data, const size_t size, const bool isMappable)
+    void Buffer::InitializeOnCpu(const void* data, const size_t size, const bool isMapped)
+    {
+        InitializeOnCpu(size, isMapped);
+        UpdateData(data);
+    }
+
+    void Buffer::InitializeOnCpu(const size_t size, const bool isMapped)
     {
         BufferInfo bufferInfo{};
 
@@ -35,14 +96,50 @@ namespace VkCore
         bufferInfo.m_AllocCreateFlags = VMA_ALLOCATION_CREATE_HOST_ACCESS_SEQUENTIAL_WRITE_BIT;
         bufferInfo.m_MemoryUsage = VMA_MEMORY_USAGE_AUTO_PREFER_HOST;
 
-        if (isMappable)
+        if (isMapped)
         {
             bufferInfo.m_AllocCreateFlags |= VMA_ALLOCATION_CREATE_MAPPED_BIT;
         }
 
-        ServiceLocator::GetAllocatorService().CreateBuffer(bufferInfo, m_Allocation, &m_AllocationInfo);
+        m_Buffer = ServiceLocator::GetAllocatorService().CreateBuffer(bufferInfo, m_Allocation, &m_AllocationInfo);
 
-        std::memcpy(m_AllocationInfo.pMappedData, data, size);
+        if (isMapped && m_AllocationInfo.pMappedData == nullptr)
+        {
+            const char* errMsg = "Buffer was mapped, but the pMappedData is nullptr!";
+            LOG(Vulkan, Fatal, errMsg)
+            throw std::runtime_error(errMsg);
+        }
+
+        m_Size = size;
+        m_IsMapped = isMapped;
+        m_IsHostVisible = true;
+    }
+
+    void Buffer::UpdateData(const void* data)
+    {
+        UpdateData(data, m_Size);
+    }
+
+    void Buffer::UpdateData(const void* data, const size_t size)
+    {
+
+        if (size == 0 || data == nullptr)
+        {
+            LOG(Vulkan, Warning, "Data wasn't updated! It is either NULL or the size of the data is 0!")
+            return;
+        }
+
+        if (m_IsMapped)
+        {
+            std::memcpy(m_AllocationInfo.pMappedData, data, size);
+            return;
+        }
+
+        void* mappedPtr = nullptr;
+
+        ServiceLocator::GetAllocatorService().MapMemory(m_Allocation, mappedPtr);
+        std::memcpy(mappedPtr, data, size);
+        ServiceLocator::GetAllocatorService().UnmapMemory(m_Allocation);
     }
 
     vk::BufferUsageFlags Buffer::GetUsageFlags() const
@@ -92,6 +189,7 @@ namespace VkCore
     void Buffer::Destroy()
     {
         ServiceLocator::GetAllocatorService().DestroyBuffer(*this);
+        m_Buffer = VK_NULL_HANDLE;
     }
 
 } // namespace VkCore
